Move tests from test_worker.py to test_gpu_worker.py
Review comment requested this:

- https://github.com/vllm-project/vllm/pull/45053#discussion_r3453324801
